lots of shit; fixed a shit load of bugs, added pages.

+++ index.js
@@ -2,23 +2,20 @@ require("dotenv").config();
 
 // depedency
 const fs = require("fs");
-const {Client, Intents, Collection, MessageEmbed} = require('discord.js');
+const {Client, Intents, Collection, MessageEmbed, Permissions} = require('discord.js');
 const mysql = require('mysql');
 const http = require('http');
-const Logger = require("./exports/logger");
+const { LogTypes, Log } = require("./exports/logger");
 const { SQL } = require("./exports/databaseManager");
 const { Config } = require('./exports/config');
-const { count } = require("console");
+const { BaldGaming } = require("./exports/baldgaming");
+const { config } = require("dotenv");
 
 var db;
 
 // vars used by the client
 const client = new Client({intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_MESSAGES, Intents.FLAGS.GUILD_MEMBERS, Intents.FLAGS.GUILD_BANS, Intents.FLAGS.DIRECT_MESSAGES]});
 client.commands = new Collection();
-
-client.ranks = {
-    
-}
 
 client.EmbedMessage = function(title, fields, user, thumbnail, desc){
 	if(!thumbnail){
@@ -87,32 +84,32 @@ client.isStaff = function(minrank, member){
 	return false;
 }
 
+client.hasAccess = function(member, accessLevel) {
+    if(member.permissions.has(Permissions.FLAGS.ADMINISTRATOR, false))
+        return true;
+
+    // probably not the best way to do this; hell, its 100% not, but whatever
+    if(typeof accessLevel == "object") {
+        accessLevel.forEach(val => {
+            if(client.hasRole(member, val)) {
+                return true;
+            }
+        });
+    }
+
+    if(typeof accessLevel === "bigint" && member.permissions.has(accessLevel)) {
+        return true;
+    }
+
+    return false; // default to false to prevent any unwanted pregnancy.
+}
+
 
 client.on('ready', () => {
-	console.log("Creating database object...");
+    Log(LogTypes.INFO, "Booting up Bald Bot!");
 
-	console.log(`Logged in as ${client.user.tag}!`);
+	Log(LogTypes.INFO, `Logged in as ${client.user.tag}!`);
 });
-
-client.on("guildMemberAdd", (member) => {
-	console.log(`[New Member] Checking if ${member.name} is verified.`);
-
-    SQL.Query("SELECT `identifier`, `discord` FROM `users` WHERE `discord` = 'discord:?'", [member.id], (success, res) => {
-        if(!success) {
-            console.error(`[FATAL] An error occured while attempting to check ${member.name}'s verification. Error: ${res}`);
-            return;
-        }
-        res = res['result'];
-
-        if(res.length > 0) {
-            console.log(`[INFO] ${member.name} (${member.id}) is verified on FiveM with account ${res.identifier}`);
-
-            member.roles.add("735631264153075814");
-            member.user.send("Welcome back! You've previously linked your discord with us, and have been granted the verified role automatically.");
-        }
-    });
-});
-
 
 // Load Commands
 const loadCommands = fs.readdirSync('./commands');
@@ -127,12 +124,23 @@ for(const folder of loadCommands){
 }
 
 client.on('messageCreate', msg => {
+    if(msg.author == client.user) // prevent the bot from saying something retarded and crashing down here. ie on the filter, it says the word the user tried to use.
+        return;
+    
     if(msg.content.indexOf("stearncommunytiy.ru") != -1 && !msg.member.user.bot) {
         msg.member.send("This link is a known phishing site. You have been banned from the server, and your message as been removed.");
         msg.member.ban({days: 7, reason: "Posting phishing links."});
         msg.delete();
 
         return;
+    } else {
+        Config.FilteredWords.forEach(val => {
+            if(msg.content.indexOf(val) != -1 && (!msg.member.user.bot && !client.hasAccess(msg.member, BaldGaming.AccessLevels.mod))) {
+                msg.member.send(`Wow, that word is so fucking rude. Shut your ass up. (You said a word in the filter list: '${val}')`);
+
+                msg.delete();
+            }
+        });
     }
 
 	if(msg.content.substring(0, 1)== "!"){
@@ -147,29 +155,31 @@ client.on('messageCreate', msg => {
 		let argtxt = args.join(" ");
 
 		if(args.length == 0) argtxt = "<NONE>"
-		console.log(msg.member.user.username + " attempted to run command " + cmd + " with args " + argtxt);
+		Log(LogTypes.ERROR, msg.member.user.username + " attempted to run command " + cmd + " with args " + argtxt);
 
 		if(cmdobj.permission == 0){
 			if(args.length < cmdobj.reqargs){
-				msg.reply("you are missing required args! Usage: ```!" + cmd + " " + cmdobj.usage + "```");
+				msg.reply("You're missing required args! Usage: ```!" + cmd + " " + cmdobj.usage + "```");
 				return;
 			}
 
 			try {
 				cmdobj.execute(msg, args)
 			} catch(error){
-				console.log(msg.member.user.username + " had a fatal error while running " + cmd + " error: \n" + error);
+				Log(LogTypes.ERROR, msg.member.user.username + " had a fatal error while running " + cmd + " error: \n" + error);
+
 				msg.reply("Error running command!\n```" + error + "```");
 			}
-		}else if(msg.member.hasPermission(cmdobj.permission) || msg.member.hasPermission('ADMINISTRATOR')){
+		}else if(client.hasAccess(msg.member, cmdobj.permission)){
 			if(args.length < cmdobj.reqargs){
-				msg.reply("you are missing required args! Usage: ```!" + cmd + " " + cmdobj.usage + "```");
+				msg.reply("You're missing required args! Usage: ```!" + cmd + " " + cmdobj.usage + "```");
+
 				return;
 			}
 			try {
 				cmdobj.execute(msg, args)
 			} catch(error){
-				console.log(msg.member.user.username + " had a fatal error while running " + cmd + " error: \n" + error);
+				Log(LogTypes.ERROR, msg.member.user.username + " had a fatal error while running " + cmd + " error: \n" + error);
 	
 				msg.reply("Error running command!\n```" + error + "```");
 			}
